Initial unified Samsung power HAL
This is a power HAL for non-qcom Samsung devices.

It currently supports standard multi-core systems as well as
big.LITTLE configurations.

This HAL auto-detects your touchscreen and touchkey input switches
and handles the interactive state requests.

I also implemented a check to make sure your touchkeys won't turn
on when they're not supposed to (for example when a CMHW class disables
the capacitive buttons and activates a software navigation solution).
This previously needed a kernel hack to make sure the touchkeys wouldn't
be activated when resuming from suspend.

Change-Id: I22888c5ac490c85c417c5457ea4f457f78ed3521

Conflicts:
	Android.mk